+++ lib/src/test/java/com/smileidentity/viewmodel/document/DocumentCaptureViewModelTest.kt
@@ -4,6 +4,7 @@ import androidx.camera.view.CameraController
 import com.google.mlkit.vision.objects.ObjectDetector
 import com.smileidentity.SmileID
 import com.smileidentity.compose.document.DocumentCaptureSide
+import com.smileidentity.models.AutoCapture
 import com.ujizin.camposer.state.CameraState
 import com.ujizin.camposer.state.ImageCaptureResult
 import io.mockk.Runs
@@ -11,17 +12,17 @@ import io.mockk.every
 import io.mockk.just
 import io.mockk.mockk
 import io.mockk.slot
-import java.io.File
 import junit.framework.TestCase.assertEquals
 import junit.framework.TestCase.assertFalse
 import junit.framework.TestCase.assertNull
 import junit.framework.TestCase.assertTrue
-import kotlin.time.Duration.Companion.seconds
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.test.setMain
 import org.junit.Before
 import org.junit.Test
+import java.io.File
+import kotlin.time.Duration.Companion.seconds
 
 @OptIn(ExperimentalCoroutinesApi::class)
 class DocumentCaptureViewModelTest {
@@ -38,7 +39,7 @@ class DocumentCaptureViewModelTest {
             knownAspectRatio = null,
             objectDetector = objectDetector,
             autoCaptureTimeout = 10.seconds,
-            enableAutoCapture = false,
+            autoCapture = AutoCapture.AutoCapture,
             metadata = mutableListOf(),
         )
     }