+++ lib/src/main/java/com/smileidentity/metadata/models/MetadataModels.kt
@@ -24,4 +24,6 @@ enum class WrapperSdkName(val value: String) {
     Flutter("flutter"),
 
     ReactNative("react_native"),
+
+    ReactNativeExpo("react_native_expo"),
 }