+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 
 * Removed `AntiFraud` response in `JobStatus` calls
 * Updated the `targetSdk` to 36 and updated the AGP version
+* Removed the default `ConsentInformation`
 
 ## 11.0.5 - July 2, 2025