+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Release Notes
 
-# Unreleased
+# 11.1.0
 
 ### Added